Fix upsample shape infer bug continue (#8159)
* fix_upsample_shape_infer_bug

* fix 5 nearest

* add 5 nearest test

* fix 5 nearest test

* fix 1 linear

* fix 4 bilinear

* fix 4 bicubic

* modify bicubic 2d file name

* fix 5 trilinear

* fix exception info

* fix exception info

* fix bug

* modify interpolate

* change float to double

* rm useless SI64ArrayAttr: in OneFlowUserOps

* rm useless import in cpp

* update

* add judge for output_size

* update oneflow/oneflow/core/autograd/gradient_funcs/upsample.cpp

* add grad in td

* test failed

* fix small failed case in upsample

* solve test error

* change float to double

* align to fix_upsample_shape_infer_bug

* align to fix_upsample_shape_infer_bug

+++ oneflow/user/kernels/upsample_kernel.cpp
@@ -114,8 +114,8 @@ class UpsampleNearestCPUKernel final : public user_op::OpKernel {
   void Compute(user_op::KernelComputeContext* ctx) const override {
     const user_op::Tensor* x_blob = ctx->Tensor4ArgNameAndIndex("x", 0);
     user_op::Tensor* y_blob = ctx->Tensor4ArgNameAndIndex("y", 0);
-    const float height_scale = ctx->Attr<float>("height_scale");
-    const float width_scale = ctx->Attr<float>("width_scale");
+    const float height_scale = ctx->Attr<double>("height_scale");
+    const float width_scale = ctx->Attr<double>("width_scale");
     const int64_t elem_cnt = y_blob->shape().elem_cnt();
     NdIndexOffsetHelper<int64_t, 4> in_helper(x_blob->shape().At(0), x_blob->shape().At(1),
                                               x_blob->shape().At(2), x_blob->shape().At(3));
@@ -141,8 +141,8 @@ class UpsampleNearestGradCPUKernel final : public user_op::OpKernel {
     Memset<DeviceType::kCPU>(ctx->stream(), dx_blob->mut_dptr<T>(), 0,
                              dx_blob->shape().elem_cnt() * sizeof(T));
     const user_op::Tensor* dy_blob = ctx->Tensor4ArgNameAndIndex("dy", 0);
-    const float height_scale = ctx->Attr<float>("height_scale");
-    const float width_scale = ctx->Attr<float>("width_scale");
+    const float height_scale = ctx->Attr<double>("height_scale");
+    const float width_scale = ctx->Attr<double>("width_scale");
     const int64_t elem_cnt = dy_blob->shape().elem_cnt();
     NdIndexOffsetHelper<int64_t, 4> dy_helper(dy_blob->shape().At(0), dy_blob->shape().At(1),
                                               dy_blob->shape().At(2), dy_blob->shape().At(3));
@@ -182,8 +182,8 @@ class UpsampleBilinearCPUKernel final : public user_op::OpKernel {
   void Compute(user_op::KernelComputeContext* ctx) const override {
     const user_op::Tensor* x_blob = ctx->Tensor4ArgNameAndIndex("x", 0);
     user_op::Tensor* y_blob = ctx->Tensor4ArgNameAndIndex("y", 0);
-    const float height_scale = ctx->Attr<float>("height_scale");
-    const float width_scale = ctx->Attr<float>("width_scale");
+    const float height_scale = ctx->Attr<double>("height_scale");
+    const float width_scale = ctx->Attr<double>("width_scale");
     const bool align_corners = ctx->Attr<bool>("align_corners");
     const int64_t elem_cnt = y_blob->shape().elem_cnt();
     NdIndexOffsetHelper<int64_t, 4> in_helper(x_blob->shape().At(0), x_blob->shape().At(1),
@@ -217,8 +217,8 @@ class UpsampleBilinearGradCPUKernel final : public user_op::OpKernel {
     Memset<DeviceType::kCPU>(ctx->stream(), dx_blob->mut_dptr<T>(), 0,
                              dx_blob->shape().elem_cnt() * sizeof(T));
     const user_op::Tensor* dy_blob = ctx->Tensor4ArgNameAndIndex("dy", 0);
-    const float height_scale = ctx->Attr<float>("height_scale");
-    const float width_scale = ctx->Attr<float>("width_scale");
+    const float height_scale = ctx->Attr<double>("height_scale");
+    const float width_scale = ctx->Attr<double>("width_scale");
     const bool align_corners = ctx->Attr<bool>("align_corners");
     const int64_t elem_cnt = dy_blob->shape().elem_cnt();
     NdIndexOffsetHelper<int64_t, 4> dy_helper(dy_blob->shape().At(0), dy_blob->shape().At(1),
